// apps/tizen/src/lib/content/stores/content/types.ts
export type SortBy = 'name' | 'date' | 'recent'
export type SortOrder = 'asc' | 'desc'
export type GroupBy = 'none' | 'group' | 'year' | 'alphabetic'

export interface LayoutData {
  categoryBrowser: number
  contentBrowser: number
}

export interface PlayerData {
  sortBy: SortBy
  sortOrder: SortOrder
  groupBy: GroupBy
}

export interface UserData {
  watchables: Record<string, UserItemData>
  hiddenGroups: string[]
  stickyGroups: string[]
  playerData: PlayerData
  layoutData: LayoutData
}